Gl/add config logic

+++ src/BaseConfig.cpp
@@ -146,7 +146,7 @@ void    BaseConfig::set_redirect(bool has_separators, const std::string &content
 const std::string&              BaseConfig::get_root( void ) const{return (this->_root);}
 AutoBool                        BaseConfig::get_autoindex( void ) const{return (this->_autoindex);}
 const ErrorPage&                BaseConfig::get_error_pages( void ) const{return (this->_error_pages);}
-int                             BaseConfig::get_max_body_size( void ) const {return (this->_client_max_body_size);}
+long long                       BaseConfig::get_max_body_size( void ) const {return (this->_client_max_body_size);}
 const std::vector<std::string>& BaseConfig::get_indexes( void ) const{return (this->_indexes);}
 const std::vector<Redirect>&    BaseConfig::get_redirect( void ){return (this->_redirect);}
